refactor: popup handling stuff
