+++ src/lib/ui/components/window/WindowContainer.svelte
